requirements.txt
fastapi
uvicorn
motor>=3.1.1
openai
pydantic
python-dotenv
razorpay
python-multipart
certifi>=2023.7.22
twilio>=8.0.0
redis>=4.5.5
langdetect
